fix ntp sync status check daemon
